refactor: divide the code into modules
